LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/video-poster.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""


FAIL Speculative parsing, page load: video-poster Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
